Remove colors from default PICKDATE_SPANATTRS
Colors can be applied using config.rasi, for example:

message, textbox {
    background-color: #285577;
    text-color: #ffffff;
}

On the other hand, text styles (bold, italic, etc.) can only be
applied to 'element-text' widgets with the 'highlight' attribute,
and the message box doesn't contain any 'element-text' widgets.
So AFAICT it's not possible to make the message box bold using
config.rasi, and PICKDATE_SPANATTRS is still needed for that.